Bug fixes and enhancements
o Fix Issue 85 - camera worker thread hang

 - worker threads should be doing CoInitializeEx(COINIT_MULTITRHEADED)
   since they do not have a windows message loop

 - it is not safe for worker threads to directly reference camera COM
   interface pointers instantiated in the main thread. Use the
   GlobalInterfaceTable to marshall the pointers just like we do with
   the ASCOM Mount interface pointers.

 - make sure that when worker threads call Yield they do not try to
   pump windows messages. Eventually we should remove all the Yield
   calls from code that runs in worker threads, but for now the Yield
   calls are still there but have no effect in the worker threads.

 - this change also Updates Issue 87: the ASCOM camera simulator no
   longer crashes PHD2 (it still does not work yet though)

o remove a bunch of casting by having
  pFrame->RequestedExposureDuration() return int

o AO Secondary mount bumping enhancements

    - always start a new bump when the AO is outside the safe range,
      rather than waiting for the current bump to complete

    - dynamically adjust the minumim bump increment so that bumping
      becomes more aggressive as long as the AO position is remaining
      outside the safe range

    - keep track of the recent average position of the AO and base the
      bump on the average position, not the current position

    - on the AO graph, display the average position, the current bump
      vector and the remaining bump vector

o fix an issue where an error during camera capture would casue PHD2
  to get stuck with "waiting for devices to stop"

+++ trunk/cam_INovaPLC.cpp
@@ -67,8 +67,8 @@ bool Camera_INovaPLCClass::Connect() {
 void Camera_INovaPLCClass::InitCapture() {
     // Run after any exp change / at the start of a loop
     DS_CAMERA_STATUS rval;
-    double ExpDur = pFrame->RequestedExposureDuration();
-    int exp_lines = (int) ExpDur * 1000 / RowTime;
+    int ExpDur = pFrame->RequestedExposureDuration();
+    int exp_lines = ExpDur * 1000 / RowTime;
     rval = DSCameraSetExposureTime(exp_lines);
     wxMilliSleep(100);
     if (rval != STATUS_OK)
@@ -77,10 +77,7 @@ void Camera_INovaPLCClass::InitCapture() {
     wxMilliSleep(100);
     if (rval != STATUS_OK)
         wxMessageBox(_T("Error setting gain"), _("Error"));
-
-
 }
-
 
 bool Camera_INovaPLCClass::PulseGuideScope(int direction, int duration) {
     unsigned char dircode = 0;
@@ -126,7 +123,7 @@ bool Camera_INovaPLCClass::Capture(int duration, usImage& img, wxRect subframe, 
         Disconnect();
         return true;
     }
-    int ExpDur = (int) pFrame-> RequestedExposureDuration();
+    int ExpDur = pFrame->RequestedExposureDuration();
 
     if (duration != ExpDur) { // reset the exp time - and pause -- we have had a change here from the current value
         rval = DSCameraSetExposureTime(ExpDur * 1000 / RowTime);
