+++ trunk/worker_thread.h
@@ -97,7 +97,7 @@ protected:
 
     /*************      Expose      **************************/
 public:
-    void EnqueueWorkerThreadExposeRequest(usImage *pImage, double exposureDuration, const wxRect& subframe);
+    void EnqueueWorkerThreadExposeRequest(usImage *pImage, int exposureDuration, const wxRect& subframe);
 protected:
     bool HandleExpose(MyFrame::EXPOSE_REQUEST *pArgs);
     void SendWorkerThreadExposeComplete(usImage *pImage, bool bError);
